+++ src/event/intermediate_throw_event.rs
@@ -120,49 +120,14 @@ impl Stream for IntermediateThrowEvent {
 
 #[cfg(test)]
 mod tests {
-    use crate::bpmn::schema::*;
+    use crate::bpmn::parse;
     use crate::event::ProcessEvent;
     use crate::model;
     use crate::test::Mailbox;
 
     #[tokio::test]
     async fn throw_none_event() {
-        let mut definitions = Definitions {
-            root_elements: vec![Process {
-                id: Some("proc1".into()),
-                flow_elements: vec![
-                    StartEvent {
-                        id: Some("start".into()),
-                        ..Default::default()
-                    }
-                    .into(),
-                    IntermediateThrowEvent {
-                        id: Some("throw".into()),
-                        ..Default::default()
-                    }
-                    .into(),
-                    EndEvent {
-                        id: Some("end".into()),
-                        ..Default::default()
-                    }
-                    .into(),
-                ],
-                ..Default::default()
-            }
-            .into()],
-            ..Default::default()
-        };
-
-        definitions
-            .find_by_id_mut("proc1")
-            .unwrap()
-            .downcast_mut::<Process>()
-            .unwrap()
-            .establish_sequence_flow("start", "throw", "s1", None::<FormalExpression>)
-            .unwrap()
-            .establish_sequence_flow("throw", "end", "s2", None::<FormalExpression>)
-            .unwrap();
-
+        let definitions = parse(include_str!("test_models/throw_none_event.bpmn")).unwrap();
         let model = model::Model::new(definitions).spawn().await;
 
         let handle = model.processes().await.unwrap().pop().unwrap();
@@ -179,54 +144,7 @@ mod tests {
 
     #[tokio::test]
     async fn throw_signal_event() {
-        let mut definitions = Definitions {
-            root_elements: vec![
-                Process {
-                    id: Some("proc1".into()),
-                    flow_elements: vec![
-                        StartEvent {
-                            id: Some("start".into()),
-                            ..Default::default()
-                        }
-                        .into(),
-                        IntermediateThrowEvent {
-                            id: Some("throw".into()),
-                            event_definitions: vec![SignalEventDefinition {
-                                signal_ref: Some("sig1".into()),
-                                ..Default::default()
-                            }
-                            .into()],
-                            ..Default::default()
-                        }
-                        .into(),
-                        EndEvent {
-                            id: Some("end".into()),
-                            ..Default::default()
-                        }
-                        .into(),
-                    ],
-                    ..Default::default()
-                }
-                .into(),
-                Signal {
-                    id: Some("sig1".into()),
-                    ..Default::default()
-                }
-                .into(),
-            ],
-            ..Default::default()
-        };
-
-        definitions
-            .find_by_id_mut("proc1")
-            .unwrap()
-            .downcast_mut::<Process>()
-            .unwrap()
-            .establish_sequence_flow("start", "throw", "s1", None::<FormalExpression>)
-            .unwrap()
-            .establish_sequence_flow("throw", "end", "s2", None::<FormalExpression>)
-            .unwrap();
-
+        let definitions = parse(include_str!("test_models/throw_signal_event.bpmn")).unwrap();
         let model = model::Model::new(definitions).spawn().await;
 
         let handle = model.processes().await.unwrap().pop().unwrap();